Move tooltip to its own component
