+++ webview/src/experiments/components/cell/Cell.tsx
@@ -1,31 +1,13 @@
 import { Experiment } from 'dvc/src/experiments/webview/contract'
 import React from 'react'
 import { Cell } from 'react-table'
+import { CellTooltip } from './CellTooltip'
 import { CopyButton } from '../copyButton/CopyButton'
 import Tooltip, {
   CELL_TOOLTIP_DELAY
 } from '../../../shared/components/tooltip/Tooltip'
 import { formatFloat } from '../../util/numberFormatting'
 import styles from '../table/styles.module.scss'
-
-const groupLabels: Record<string, string> = {
-  metrics: 'Metric',
-  params: 'Parameter'
-}
-
-const CellTooltip: React.FC<{
-  cell: Cell<Experiment, string | number>
-}> = ({ cell }) => {
-  const {
-    column: { group },
-    value
-  } = cell
-  return (
-    <>
-      {groupLabels[group as string]}: {value}
-    </>
-  )
-}
 
 const UndefinedCell = (
   <div className={styles.innerCell}>
